+++ raiden_contracts/data_0.10.0/source/test/TokenNetworkInternalsTest.sol
@@ -227,11 +227,11 @@ contract TokenNetworkSignatureTest is TokenNetwork {
         );
     } */
 
-    /* function recoverAddressFromWithdrawMessagePublic(
+    function recoverAddressFromWithdrawMessagePublic(
         uint256 channel_identifier,
         address participant,
         uint256 amount_to_withdraw,
-        bytes signature
+        bytes memory signature
     )
         view
         public
@@ -243,7 +243,7 @@ contract TokenNetworkSignatureTest is TokenNetwork {
             amount_to_withdraw,
             signature
         );
-    } */
+    }
 }
 
 contract TokenNetworkUtilsTest is TokenNetwork {